pyproject.toml
[tool.poetry]
name = "moody"
version = "0.1.0"
description = ""
authors = ["Danimal <chao-mu@hackpoetic.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.2"
Flask-SQLAlchemy = "^2.5.1"
PyYAML = "^6.0"
psycopg2 = "^2.9.3"
uWSGI = "^2.0.20"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
